refactor: 마이페이지 UI 분리
