+++ frontend/src/components/common/PageTitle.tsx
@@ -1,0 +1,13 @@
+interface TitleProps {
+  title: string;
+}
+
+const PageTitle = ({ title }: TitleProps) => {
+  return (
+    <h1 className={"text-bold-l mb-6 text-gray-black dark:text-white"}>
+      {title}
+    </h1>
+  )
+}
+
+export default PageTitle;
