+++ frontend/src/components/layout/SidebarPageLayout.tsx
@@ -3,14 +3,15 @@ import Sidebar from "@components/common/Sidebar.tsx";
 
 interface SidebarPageLayoutProps {
   children: React.ReactNode;
+  childrenClassName?: string;
 }
 
-const SidebarPageLayout = ({ children }: SidebarPageLayoutProps) => {
+const SidebarPageLayout = ({ children, childrenClassName = "" }: SidebarPageLayoutProps) => {
   return (
     <section className={sectionWithSidebar}>
       <Sidebar />
-      {children}
-    </section>
+      <div className={childrenClassName}>{children}</div>
+    </section >
   );
 };
